Improve `yarn dev` DX in `block-scripts` (#317)
