+++ packages/block-scripts/package.json
@@ -1,6 +1,6 @@
 {
   "name": "block-scripts",
-  "version": "0.0.6",
+  "version": "0.0.7",
   "description": "Block development framework",
   "keywords": [
     "blockprotocol",
