+++ packages/block-template/package.json
@@ -1,6 +1,6 @@
 {
   "name": "block-template",
-  "version": "0.0.17",
+  "version": "0.0.18",
   "description": "Block template",
   "keywords": [
     "blockprotocol",
@@ -24,11 +24,11 @@
     "serve": "block-scripts serve"
   },
   "dependencies": {
-    "blockprotocol": "0.0.9"
+    "blockprotocol": "0.0.10"
   },
   "devDependencies": {
-    "block-scripts": "0.0.6",
-    "mock-block-dock": "0.0.9",
+    "block-scripts": "0.0.7",
+    "mock-block-dock": "0.0.10",
     "react": "^17.0.2",
     "react-dom": "^17.0.2"
   },
