+++ packages/mock-block-dock/package.json
@@ -1,6 +1,6 @@
 {
   "name": "mock-block-dock",
-  "version": "0.0.9",
+  "version": "0.0.10",
   "description": "A mock embedding application for Block Protocol blocks",
   "keywords": [
     "blockprotocol",
@@ -35,7 +35,7 @@
     "run-dev-server": "cross-env NODE_ENV=development webpack-dev-server --config dev/webpack.config.js"
   },
   "dependencies": {
-    "blockprotocol": "0.0.9",
+    "blockprotocol": "0.0.10",
     "lodash": "^4.17.21",
     "uuid": "^8.3.2"
   },
