+++ packages/react-block-loader/dev/webpack.config.js
@@ -41,7 +41,6 @@ module.exports = {
       ],
     },
     hot: true,
-    open: process.env.BROWSER !== "none",
     port: 9090,
     static: {
       directory: __dirname,
