+++ packages/react-block-loader/package.json
@@ -1,6 +1,6 @@
 {
   "name": "react-block-loader",
-  "version": "0.0.2",
+  "version": "0.0.3",
   "description": "A React component to load Block Protocol blocks from a URL",
   "keywords": [
     "blockprotocol",
@@ -37,7 +37,7 @@
   },
   "dependencies": {
     "@lit-labs/react": "^1.0.4",
-    "blockprotocol": "0.0.9",
+    "blockprotocol": "0.0.10",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "uuid": "^8.3.2"
@@ -56,7 +56,7 @@
     "cross-env": "^7.0.3",
     "html-webpack-plugin": "^5.5.0",
     "lodash": "^4.17.21",
-    "mock-block-dock": "0.0.9",
+    "mock-block-dock": "0.0.10",
     "regenerator-runtime": "^0.13.9",
     "rimraf": "^3.0.2",
     "twind": "^0.16.17",
