+++ site/package.json
@@ -34,7 +34,7 @@
     "ajv": "^8.9.0",
     "algoliasearch": "4.12.1",
     "axios": "^0.24.0",
-    "blockprotocol": "0.0.9",
+    "blockprotocol": "0.0.10",
     "busboy": "^1.4.0",
     "clsx": "^1.1.1",
     "connect-mongo": "^4.6.0",
@@ -56,7 +56,7 @@
     "lodash": "^4.17.21",
     "md5": "^2.3.0",
     "mime-types": "^2.1.34",
-    "mock-block-dock": "0.0.9",
+    "mock-block-dock": "0.0.10",
     "mongodb": "^4.2.2",
     "next": "^12.1.5",
     "next-connect": "^0.12.1",
